+++ BookGen.Launch/Properties/Settings.Designer.cs
@@ -37,18 +37,6 @@ namespace BookGen.Launch.Properties {
         
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("")]
-        public string FolderListJson {
-            get {
-                return ((string)(this["FolderListJson"]));
-            }
-            set {
-                this["FolderListJson"] = value;
-            }
-        }
-        
-        [global::System.Configuration.UserScopedSettingAttribute()]
-        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
         [global::System.Configuration.DefaultSettingValueAttribute("True")]
         public bool AutoExitLauncher {
             get {